move snackbar into dir
