+++ packages/wallet-sdk/src/components/Snackbar/Snackbar.scss
